+++ playground-env/src/main/scala/org/ergoplatform/playgroundenv/models/BlockchainContext.scala
@@ -2,5 +2,5 @@ package org.ergoplatform.playgroundenv.models
 
 trait BlockchainContext {}
 
-case class DummyBlockchainContext(blockchain: BlockchainSimulation)
+case class DummyBlockchainContextImpl(blockchain: BlockchainSimulation)
   extends BlockchainContext {}
